[Downloader] Manage of None client

+++ tests/test_Downloader.py
@@ -54,16 +54,17 @@ class TestDownloader(unittest.TestCase):
 		if self.testTransmission:
 			self.d = Downloader.Downloader()
 			self.d.loadConfig(self.configFileTransmission)
-			filename = "{0}/{1}".format(os.path.dirname(os.path.abspath(__file__)),'test.torrent')
+			if self.d.conf['client'] is not None:
+				filename = "{0}/{1}".format(os.path.dirname(os.path.abspath(__file__)),'test.torrent')
 			
-			tmpfile = unicode(tempfile.mkstemp('.torrent')[1])
-			os.remove(tmpfile)
-			shutil.copyfile(filename, tmpfile)
+				tmpfile = unicode(tempfile.mkstemp('.torrent')[1])
+				os.remove(tmpfile)
+				shutil.copyfile(filename, tmpfile)
 			
-			id = self.d.add_torrent(tmpfile,delTorrent=True)
-			self.assertIsInstance(id,int)
-			self.assertFalse(os.path.isfile(tmpfile))
-			return id
+				id = self.d.add_torrent(tmpfile,delTorrent=True)
+				self.assertIsInstance(id,int)
+				self.assertFalse(os.path.isfile(tmpfile))
+				return id
 			
 	def test_get_status_transmission(self):
 		if self.testTransmission:
